Cleanup + Add getEthBridgeInformation function in networks file.
